Seperate DiscordEngine from VoiceEngine, make protocols less dependant
